test: create chat-store-mocks for helper testing methods
