+++ src/chats/use-chat-automation.test.tsx
@@ -1,112 +1,10 @@
+import { createMockChatInstance, createMockUseChat, hydrateStore, resetStore } from '@/test-utils/chat-store-mocks'
 import { createQueryTestWrapper } from '@/test-utils/react-query'
 import { getClock } from '@/testing-library'
-import type { Model, ThunderboltUIMessage } from '@/types'
-import { type Chat, type useChat } from '@ai-sdk/react'
+import type { ThunderboltUIMessage } from '@/types'
 import { act, cleanup, renderHook } from '@testing-library/react'
 import { afterEach, beforeEach, describe, expect, it, mock } from 'bun:test'
 import { useChatAutomation } from './use-chat-automation'
-import { useChatStore } from './chat-store'
-
-// Mock Chat instance - minimal implementation for testing
-const createMockChatInstance = (
-  messages: ThunderboltUIMessage[] = [],
-  status: 'ready' | 'streaming' = 'ready',
-): Chat<ThunderboltUIMessage> => {
-  const sendMessage = mock((_params: { text: string; metadata?: Record<string, unknown> }) => {
-    // Mock implementation
-  })
-  const regenerate = mock(() => Promise.resolve())
-
-  return {
-    id: 'test-chat-id',
-    messages,
-    sendMessage,
-    status,
-    regenerate,
-    stop: mock(),
-    append: mock(),
-    reload: mock(),
-    setMessages: mock(),
-    setData: mock(),
-    setStatus: mock(),
-  } as unknown as Chat<ThunderboltUIMessage>
-}
-
-// Mock useChat hook that reads from chat instance
-const createMockUseChat = (chatInstance: Chat<ThunderboltUIMessage>) => {
-  return ((_options?: { chat?: Chat<ThunderboltUIMessage> }) => ({
-    id: chatInstance.id,
-    status: chatInstance.status,
-    messages: chatInstance.messages,
-    error: undefined,
-    isLoading: false,
-    reload: mock(),
-    stop: chatInstance.stop,
-    append: mock(),
-    setMessages: mock(),
-    setData: mock(),
-    sendMessage: chatInstance.sendMessage,
-    regenerate: chatInstance.regenerate,
-    resumeStream: mock(),
-    addToolResult: mock(),
-    clearError: mock(),
-  })) as unknown as typeof useChat
-}
-
-/**
- * Helper to hydrate the store with a session (replaces old hydrate method)
- */
-const hydrateStore = (state: {
-  chatInstance: Chat<ThunderboltUIMessage>
-  chatThread: null
-  id: string
-  mcpClients: never[]
-  models: Model[]
-  selectedModel: Model | null
-  triggerData: null
-}) => {
-  const store = useChatStore.getState()
-
-  // Set models first
-  store.setModels(state.models)
-
-  // Set MCP clients
-  store.setMcpClients(state.mcpClients)
-
-  // Create session with a default model if selectedModel is null
-  const defaultModel =
-    state.selectedModel ??
-    ({
-      id: 'default-model',
-      provider: 'openai',
-      name: 'Default Model',
-      model: 'gpt-4',
-      isSystem: 0,
-      enabled: 1,
-      isConfidential: 0,
-    } as Model)
-
-  store.createSession({
-    chatInstance: state.chatInstance,
-    chatThread: state.chatThread,
-    id: state.id,
-    selectedModel: defaultModel,
-    triggerData: state.triggerData,
-  })
-  store.setCurrentSessionId(state.id)
-}
-
-/**
- * Helper to reset the store (replaces old reset method)
- */
-const resetStore = () => {
-  useChatStore.setState({
-    currentSessionId: null,
-    mcpClients: [],
-    models: [],
-    sessions: new Map(),
-  })
-}
 
 describe('useChatAutomation', () => {
   let consoleErrorSpy: ReturnType<typeof mock>
