+++ src/components/chat/chat-prompt-input.test.tsx
@@ -1,102 +1,29 @@
-import { useChatStore } from '@/chats/chat-store'
 import { resetTestDatabase, setupTestDatabase, teardownTestDatabase } from '@/dal/test-utils'
+import {
+  createMockChatInstance,
+  createMockChatThread,
+  createMockModel,
+  createMockUseChat,
+  hydrateStore,
+  resetStore,
+} from '@/test-utils/chat-store-mocks'
 import { createQueryTestWrapper } from '@/test-utils/react-query'
-import type { ChatThread, Model, ThunderboltUIMessage } from '@/types'
-import { type Chat } from '@ai-sdk/react'
+import type { Model } from '@/types'
 import { act, cleanup, render, screen } from '@testing-library/react'
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it, mock } from 'bun:test'
 import { createElement } from 'react'
 import { BrowserRouter } from 'react-router'
 import { ChatPromptInput, type ChatPromptInputRef } from './chat-prompt-input'
 
-// Mock Chat instance - minimal implementation for testing
-const createMockChatInstance = (
-  messages: ThunderboltUIMessage[] = [],
-  status: 'ready' | 'streaming' = 'ready',
-): Chat<ThunderboltUIMessage> => {
-  const sendMessage = mock((_params: { text: string; metadata?: Record<string, unknown> }) => {
-    // Mock implementation
-  })
-  const regenerate = mock(() => Promise.resolve())
-
-  return {
-    id: 'test-chat-id',
-    messages,
-    sendMessage,
-    status,
-    regenerate,
-    stop: mock(),
-    append: mock(),
-    reload: mock(),
-    setMessages: mock(),
-    setData: mock(),
-    setStatus: mock(),
-  } as unknown as Chat<ThunderboltUIMessage>
-}
-
-// Mock useChat hook
-const createMockUseChat = (chatInstance: Chat<ThunderboltUIMessage>) => {
-  return ((_options?: { chat?: Chat<ThunderboltUIMessage> }) => ({
-    id: chatInstance.id,
-    status: chatInstance.status,
-    messages: chatInstance.messages,
-    error: undefined,
-    isLoading: false,
-    reload: mock(),
-    stop: chatInstance.stop,
-    append: mock(),
-    setMessages: mock(),
-    setData: mock(),
-    sendMessage: chatInstance.sendMessage,
-    regenerate: chatInstance.regenerate,
-    resumeStream: mock(),
-    addToolResult: mock(),
-    clearError: mock(),
-  })) as unknown as typeof import('@ai-sdk/react').useChat
-}
-
-const createMockChatThread = (overrides?: Partial<ChatThread>): ChatThread => {
-  return {
-    id: 'thread-1',
-    title: 'Test Thread',
-    isEncrypted: 0,
-    ...overrides,
-  } as ChatThread
-}
-
-const createMockModel = (overrides?: Partial<Model>): Model => {
-  return {
-    id: 'model-1',
-    provider: 'thunderbolt',
-    name: 'Test Model',
-    model: 'gpt-oss-120b',
-    isSystem: 0,
-    enabled: 1,
-    isConfidential: 0,
-    contextWindow: 131072,
-    toolUsage: 1,
-    startWithReasoning: 0,
-    deletedAt: null,
-    url: null,
-    defaultHash: null,
-    apiKey: null,
-    ...overrides,
-  } as Model
-}
-
 // Mock useContextTracking hook
-const createMockUseContextTracking = (
-  isOverflowing: boolean = false,
-  isContextKnown: boolean = true,
-  usedTokens: number | null = 1000,
-  maxTokens: number | null = 2000,
-) => {
-  return (_options?: {
-    model?: Model | null
-    chatThreadId?: string
-    currentInput?: string
-    onOverflow?: () => void
-  }) => ({
+const createMockUseContextTracking =
+  (
+    isOverflowing: boolean = false,
+    isContextKnown: boolean = true,
+    usedTokens: number | null = 1000,
+    maxTokens: number | null = 2000,
+  ) =>
+  (_options?: { model?: Model | null; chatThreadId?: string; currentInput?: string; onOverflow?: () => void }) => ({
     usedTokens,
     maxTokens,
     isContextKnown,
@@ -104,11 +31,11 @@ const createMockUseContextTracking = (
     isLoading: false,
     estimateTokensForInput: (_input: string) => 0,
   })
-}
 
 // Mock useSidebar hook
-const createMockUseSidebar = (isMobile: boolean = false, openMobile: boolean = false) => {
-  return () => ({
+const createMockUseSidebar =
+  (isMobile: boolean = false, openMobile: boolean = false) =>
+  () => ({
     isMobile,
     openMobile,
     state: 'expanded' as const,
@@ -121,63 +48,6 @@ const createMockUseSidebar = (isMobile: boolean = false, openMobile: boolean = f
     isDraggingRail: false,
     setIsDraggingRail: mock(),
   })
-}
-
-/**
- * Helper to hydrate the store with a session (replaces old hydrate method)
- */
-const hydrateStore = (state: {
-  chatInstance: Chat<ThunderboltUIMessage>
-  chatThread: ChatThread | null
-  id: string
-  mcpClients: never[]
-  models: Model[]
-  selectedModel: Model | null
-  triggerData: null
-}) => {
-  const store = useChatStore.getState()
-
-  // Set models first
-  store.setModels(state.models)
-
-  // Set MCP clients
-  store.setMcpClients(state.mcpClients)
-
-  // Create session with a default model if selectedModel is null
-  const defaultModel =
-    state.selectedModel ??
-    state.models[0] ??
-    ({
-      id: 'default-model',
-      provider: 'openai',
-      name: 'Default Model',
-      model: 'gpt-4',
-      isSystem: 0,
-      enabled: 1,
-      isConfidential: 0,
-    } as Model)
-
-  store.createSession({
-    chatInstance: state.chatInstance,
-    chatThread: state.chatThread,
-    id: state.id,
-    selectedModel: defaultModel,
-    triggerData: state.triggerData,
-  })
-  store.setCurrentSessionId(state.id)
-}
-
-/**
- * Helper to reset the store (replaces old reset method)
- */
-const resetStore = () => {
-  useChatStore.setState({
-    currentSessionId: null,
-    mcpClients: [],
-    models: [],
-    sessions: new Map(),
-  })
-}
 
 /**
  * Wrapper that includes Router context for useNavigate
